# job-platform/backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MedMatch Job Platform API",
    "description": "API for job matching platform connecting medical professionals with healthcare employers. Includes application tracking, dashboard analytics, and notification system.",
    "version": "1.1.0",
    "contact": {
      "name": "MedMatch API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/candidates/{id}/applications": {
      "get": {
        "summary": "Get candidate applications",
        "description": "Get all applications for a specific candidate",
        "tags": ["Candidates", "Applications"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Candidate ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of candidate applications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "candidateId": { "type": "integer" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Application" }
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "byStatus": {
                          "type": "object",
                          "properties": {
                            "pending": { "type": "integer" },
                            "reviewed": { "type": "integer" },
                            "accepted": { "type": "integer" },
                            "rejected": { "type": "integer" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Candidate not found" }
        }
      }
    },
...
    "/dashboard/candidate/{id}": {
      "get": {
        "summary": "Candidate dashboard",
        "description": "Get dashboard statistics for a candidate",
        "tags": ["Dashboard"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Candidate ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate dashboard stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "candidateId": { "type": "integer" },
                    "candidateName": { "type": "string" },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "totalApplications": { "type": "integer" }
                      }
                    },
                    "applications": {
                      "type": "object",
                      "properties": {
                        "byStatus": {
                          "type": "object",
                          "properties": {
                            "pending": { "type": "integer" },
                            "reviewed": { "type": "integer" },
                            "accepted": { "type": "integer" },
                            "rejected": { "type": "integer" }
                          }
                        },
                        "recent": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Application" }
                        },
                        "responseRate": { "type": "integer", "description": "Percentage of applications with non-pending status" },
                        "monthlyTrend": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "date": { "type": "string" },
                              "count": { "type": "integer" }
                            }
                          }
                        }
                      }
                    },
                    "profile": {
                      "type": "object",
                      "properties": {
                        "completionPercentage": { "type": "integer" },
                        "hasCv": { "type": "boolean" },
                        "hasPortfolio": { "type": "boolean" }
                      }
                    },
                    "recommendations": {
                      "type": "object",
                      "properties": {
                        "totalJobs": { "type": "integer" },
                        "appliedTo": { "type": "integer" },
                        "newOpportunities": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Candidate not found" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Application": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "candidateId": { "type": "integer" },
          "jobId": { "type": "integer" },
          "status": { "type": "string", "enum": ["pending", "reviewed", "accepted", "rejected"] },
          "coverLetter": { "type": "string" },
          "cvUrl": { "type": "string" },
          "matchScore": { "type": "integer" },
          "appliedAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "candidateFirstName": { "type": "string" },
          "candidateLastName": { "type": "string" },
          "candidateEmail": { "type": "string" },
          "candidateSpecialty": { "type": "string" },
          "jobTitle": { "type": "string" },
          "jobSpecialty": { "type": "string" },
          "jobLocation": { "type": "string" },
          "employerName": { "type": "string" },
          "employerId": { "type": "integer" }
        }
      },
...
    }
  }
}